requirements.txt
# Core
Django==5.2.9
djangorestframework==3.16.*

# Auth (JWT)
djangorestframework-simplejwt==5.3.1

# HTTP client (Surepass)
requests==2.32.3

# CORS (Flutter app)
django-cors-headers==4.6.0

# Environment variables
python-dotenv==1.0.1

# DB driver (PostgreSQL) - optional but recommended
psycopg[binary]==3.2.3

# Utilities
pytz==2024.2

